requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-braces==1.11.0
django-cors-headers==2.1.0
django-countries==5.0
django-debug-toolbar==1.9.1
django-el-pagination==3.1.0
django-model-utils==3.0.0
django-star-ratings==0.5.6
djangorestframework==3.7.1
easy-thumbnails==2.5
idna==2.6
numpy==1.13.3
oauthlib==2.0.6
olefile==0.44
pandas==0.21.0
Pillow==4.3.0
pyflakes==1.6.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
uWSGI==2.0.15
